+++ changelog.md
@@ -2,6 +2,16 @@
 
 All notable changes to Lock - Laravel will be documented in this file. This file follows the *[Keep a CHANGELOG](http://keepachangelog.com/)* standards.
 
+## 0.1.2 - 2015-01-13
+
+### Changed
+
+- Permissions are now cached at runtime. Permissions are only reset when permissions are stored or removed.
+
+### Fixed
+
+- Fixed a bug where the guest caller had an incorrect id set
+
 ## 0.1.1 - 2014-12-17
 
 ### Fixed